handle adding and cleaning up data provider in panel editor on panel switch
